Fix bug that case the resending of validation email
When a user tries to send us a message from the website,
we first send them an email to validate their intention.

This allows us to filter out spam. There was a bug that will
continue to send the validate email to the user. This ate into our
email qouta.

This fix, checks for an existing entry in the pending validation list
before sending the email.